Added react application

+++ CollegeFootball/CollegeFootball.Domain/Entities/TeamScore.cs
@@ -13,7 +13,7 @@ namespace CollegeFootball.Domain.Entities
         public string TeamName { get; set; } = string.Empty;
         public string MascotName { get; set; } = string.Empty;
         public string LastWinDate { get; set; } = string.Empty;
-        public float? WinningPercentage { get; set; }
+        public double? WinningPercentage { get; set; }
         public int? TotalLosses { get; set; }
         public int? TotalTies { get; set; }
         public int? TotalWins { get; set; }
